requirements.txt
arcgis
requests
python-dotenv
streamlit
folium
streamlit-folium
beautifulsoup4
duckdb>=1.0.0
pandas>=2.0.0
tabulate
